requirements.txt
python-nmap
dnspython
python-whois
